outsourced many things
