+++ src/pages/Plus.js
@@ -5,14 +5,18 @@ import FooterNavigation from '../components/FooterNavigation';
 
 
 
-export default function Plus({onAddHabit}) {
+export default function Plus({
+  onAddHabit,
+  setHabitToEdit,
+  onNavigate
+}) {
 
 
 
   return (
     <>
      <h1> This is the plus page! </h1>
-     <HabitForm onAddHabit={onAddHabit} />
+     <HabitForm onAddHabit={onAddHabit} setHabitToEdit={setHabitToEdit}/>
      <FooterNavigation />
     </>
   );
